Revert "Update to CiviCRM 6.11rc"
Revert until we can plan deployment as it is a slow run

This reverts commit 5564cf521459fd3d122a5b3cc24df4e74b0f4dc0.

+++ core/CRM/Utils/Chart.php
@@ -99,7 +99,7 @@ class CRM_Utils_Chart {
     foreach ($params['multiValues'] as $i => $dataSet) {
       $output['values'][$i] = [];
       foreach ($dataSet as $k => $v) {
-        $output['values'][$i][] = ['label' => $k, 'value' => (float) $v];
+        $output['values'][$i][] = ['label' => $k, 'value' => (double) $v];
       }
     }
     if (!$output['values']) {
